Add unit test for node profiles and migrate misc tests
